Implement Tom's recommended changes for tools directory
