+++ holopy/core/tools/__init__.py
@@ -27,6 +27,7 @@ reconstructions.
 '''
 
 
-from .img_proc import normalize, detrend, zero_filter, subimage, resize, add_noise
+from .img_proc import normalize, detrend, zero_filter, subimage, resize, add_noise, fft, ifft
 from .centerfinder import center_find, centered_subimage
-from .fourier import fft, ifft
+from .utilities import _ensure_array, ensure_listlike, _ensure_pair, ensure_3d, mkdir_p
+from .utilities import coord_grid, dict_without, is_none, updated, squeeze
