+++ holopy/core/tools/img_proc.py
@@ -28,8 +28,9 @@ or detrending
 
 from ..errors import BadImage
 from .math import simulate_noise
-from .helpers import is_none, ensure_3d
+from .utilities import is_none, ensure_3d
 from scipy.signal import detrend
+from scipy import fftpack
 #from ..marray import Image, Volume, RegularGridSchema
 import numpy as np
 
@@ -234,3 +235,93 @@ def _checked_cut(arr, extent):
             raise IndexError
 
     return arr[extent].copy()
+
+
+def fft(a, overwrite=False, shift=True):
+    """
+    More convenient Fast Fourier Transform
+
+    An easier to use fft function, it will pick the correct fft to do
+    based on the shape of the Marray, and do the fftshift for you.  This
+    is intended for working with images, and thus for dimensions
+    greater than 2 does slicewise transforms of each "image" in a
+    multidimensional stack
+
+    Parameters
+    ----------
+    a : ndarray
+       The array to transform
+    overwrite : bool
+       Allow this function to overwrite the Marry you pass in.  This
+       may improve performance slightly.  Default is not to overwrite
+    shift : bool
+       Whether to preform an fftshift on the Marry to give low
+       frequences near the center as you probably expect.  Default is
+       to do the fftshift.
+
+    Returns
+    -------
+    fta : ndarray
+       The fourier transform of `a`
+    """
+    from ..marray import Marray, arr_like
+    if a.ndim is 1:
+        if shift:
+            res = fftpack.fftshift(fftpack.fft(a, overwrite_x=overwrite))
+        else:
+            res = fftpack.fft(a, overwrite_x=overwrite)
+    else:
+        if shift:
+            res = fftpack.fftshift(fftpack.fft2(a, axes=[0, 1],
+                                                 overwrite_x=overwrite),
+                                    axes=[0,1])
+        else:
+            res = fftpack.fft2(a, axes=[0, 1], overwrite_x=overwrite)
+    if isinstance(a, Marray):
+        res = arr_like(res, a)
+    return res
+
+
+def ifft(a, overwrite=False, shift=True):
+    """
+    More convenient Inverse Fast Fourier Transform
+
+    An easier to use ifft function, it will pick the correct ifft to
+    do based on the shape of the Marry, and do the fftshift for you.
+    This is indendended for working with images, and thus for
+    dimensions greater than 2 does slicewise transforms of each
+    "image" in a multidimensional stack
+
+    Parameters
+    ----------
+    a : ndarray
+       The array to transform
+    overwrite : bool
+       Allow this function to overwrite the Marry you pass in.  This
+       may improve performance slightly.  Default is not to overwrite
+    shift : bool
+       Whether to preform an fftshift on the Marry to give low
+       frequences near the center as you probably expect.  Default is to
+       do the fftshift.
+
+    Returns
+    -------
+    ifta : ndarray
+       The inverse fourier transform of `a`
+    """
+    from ..marray import Marray, arr_like
+    if a.ndim is 1:
+        if shift:
+            res = fftpack.ifft(fftpack.fftshift(a, overwrite_x=overwrite))
+        else:
+            res = fftpack.ifft(a, overwrite_x=overwrite)
+    else:
+        if shift:
+            res = fftpack.ifft2(fftpack.fftshift(a, axes=[0,1]), axes=[0, 1],
+                                 overwrite_x=overwrite)
+        else:
+            res = fftpack.ifft2(a, overwrite_x=overwrite)
+    if isinstance(a, Marray):
+        res = arr_like(res, a)
+    return res
+
