+++ holopy/propagation/convolution_propagation.py
@@ -25,8 +25,7 @@ Code to propagate objects/waves using scattering models.
 
 
 import numpy as np
-from ..core.tools.fourier import fft, ifft
-from ..core.tools.helpers import _ensure_pair, _ensure_array
+from ..core.tools import fft, ifft, _ensure_pair, _ensure_array
 from ..core import Volume, Image, Grid, UnevenGrid, VolumeSchema, Marray
 from ..core.marray import VectorGrid, dict_without, arr_like
 from ..core.metadata import interpret_args
