+++ holopy/scattering/scatterer/sphere_builtin.py
@@ -27,7 +27,7 @@ import numpy as np
 from .scatterer import CenteredScatterer, Indicators
 from ..errors import ScattererDefinitionError
 from copy import copy
-from ...core.tools.helpers import _ensure_array
+from ...core.tools import _ensure_array
 
 class Sphere_builtin(CenteredScatterer):
     '''
